services/browser-service/pyproject.toml
[project]
name = "browser-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.11.9",
    "fastapi>=0.129.0",
    "google-genai>=1.64.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.41.0",
]
